Created a readme.
Definitely need to commit a few paragraphs to how I created a simple readme file and renamed the repo. Because coming up with a name is probably the most important part of any app. Even more so than putting effort and time into a project

It's totally ok if this is never finished, because it has a fantastic name now. BASICALLY 99% DONE